Add calendar

+++ db/migrate/20170718103122_create_categories.rb
@@ -3,7 +3,7 @@ class CreateCategories < ActiveRecord::Migration[5.0]
     create_table :categories do |t|
       t.text :name
       t.text :cat_type
-      t.integer :summ
+      t.money :summ
       t.integer :user_id
 
       t.timestamps
